+++ include/player.h
@@ -12,22 +12,22 @@ typedef struct {
     bool exists;
 
     Vector position;
-    double speed;
+    float speed;
     Vector velocity;
-    double speedModifier;
+    float speedModifier;
 
     BodyCircle body;
 
     bool fire;
-    double fireTimer;
-    double reloadTime;
+    float fireTimer;
+    float reloadTime;
 
     Vector model[PLAYER_MODEL_COUNT];
     float render[PLAYER_MODEL_COUNT * 2];
 } Player;
 
 void Player_Initialize();
-void Player_Update(Player* player, double dt);
+void Player_Update(Player* player, float dt);
 void Player_Render(GPU_Target* target, Player* player);
 
 Uint32 Player_Reload(Uint32 interval, void* param);
