+++ include/renderer.h
@@ -17,7 +17,7 @@ typedef struct {
 
     Vector renderCenter;
     Vector cameraPosition;
-    double zoom;
+    float zoom;
 } Renderer;
 
 extern Renderer renderer;
@@ -26,15 +26,15 @@ void Renderer_Initialize();
 void Renderer_Destroy();
 void Renderer_Reset_Camera();
 
-void Renderer_Draw_Point(GPU_Target* target, double x, double y, SDL_Color color);
+void Renderer_Draw_Point(GPU_Target* target, float x, float y, SDL_Color color);
 void Renderer_Draw_Points(GPU_Target* target, Vector points[], int pointCount, SDL_Color color);
 void Renderer_Draw_Lines(GPU_Target* target, float points[], int pointCount, SDL_Color color);
 void Renderer_Draw_Rectangle(GPU_Target* target, GPU_Rect rect, SDL_Color color, bool isFilled);
 
 void Renderer_Render();
 
-Vector Renderer_Screen_To_Game_TransformF(double screenX, double screenY, bool isRender);
-Vector Renderer_Game_To_Screen_TransformF(double gameX, double gameY, bool isRender);
+Vector Renderer_Screen_To_Game_TransformF(float screenX, float screenY, bool isRender);
+Vector Renderer_Game_To_Screen_TransformF(float gameX, float gameY, bool isRender);
 Vector Renderer_Screen_To_Game_TransformV(Vector screenPoint, bool isRender);
 Vector Renderer_Game_To_Screen_TransformV(Vector gamePosition, bool isRender);
