+++ include/world.h
@@ -29,7 +29,7 @@ extern World world;
 
 void World_Initialize();
 void World_Clean();
-void World_Update(double dt);
+void World_Update(float dt);
 void World_Physics_Update();
 void World_Render(GPU_Target* target);
